Move Test to tool package
