+++ src/main/java/com/moisesarrona/tool/TestInterface.java
@@ -1,0 +1,11 @@
+package com.moisesarrona.tool;
+
+import com.moisesarrona.segments.body.Body;
+
+public interface TestInterface {
+    public String getStrJson ();
+
+    public String getStrEdi ();
+
+    public Body getObjectEdi();
+}
